Refactored some JSON parsing
